+++ CHANGELOG.md
@@ -9,6 +9,18 @@ From v1.0.0 and on, the project will adherence strictly to Semantic Versioning.
 
 ## [Unreleased]
 
+## [0.16.0]
+
+### Added
+- "lagging" option for shrink in `co.Delay` and `co.Residual`.
+- `co.RNN`.
+- `co.LSTM`.
+- `co.GRU`.
+
+### Changed
+- `phantom_padding` renamed to `residual_shrink`.
+
+
 ## [0.15.6]
 ### Fixed
 - Missing cpp file in package.